backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.2
aiosqlite==0.19.0
pyserial==3.5
python-dotenv==1.0.0
